python/03-energy-meter-pipeline/requirements.txt
pandas
numpy
scipy
pytz
mysql-connector-python